Avoid permission errors due to file metadata copy
After the SPM batch has run on temporary files, the outputs are copied
using shutil.move. That function will by default use shutil.copy2 followed
by os.remove, when the source and destination are on different filesystems.
shutil.copy2 (as well as shutil.copy) attempt to copy file metadata (mtime,
mode, etc.) which fails if the user is not an owner of the destination file.
Using shutil.copyfile allows these calls to succeed, at the expense of not
copying the file metadata at all, which should not be an issue.